+++ CHANGELOG.md
@@ -4,9 +4,9 @@ All notable changes to this project will be documented in this file.
 ---
 ## [X.X.X]
 * Breaking change:
-	* updateNowPlayingInfo(:) was replaced by updateMetadata(:)
+	* `updateNowPlayingInfo(metadata:)` was replaced by `updateMetadata(_:)`
 * Fix:
-	* Update medatata replace current media metadata
+	* Update medatata replace current media metadata and update now playing info
 
 ## [0.14.0]
 * Features: