+++ CHANGELOG.md
@@ -2,11 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 ---
-## [X.X.X]
+## [0.15.0]
 * Breaking change:
 	* `updateNowPlayingInfo(metadata:)` was replaced by `updateMetadata(_:)`
-* Fix:
-	* Update medatata replace current media metadata and update now playing info
+* Feature:
+	* `updateMedatata(_:)` replace current media metadata and update now playing info
 
 ## [0.14.0]
 * Features: